// src/types/category.ts
export enum Category {
  ACCOMMODATION = 'accommodation',
  ACRONYMS = 'acronyms',
  AIR_TRAVEL = 'airTravel',
  ALCOHOLIC_DRINKS = 'alcoholicDrinks',
  ANIMALS = 'animals',
  BEVERAGES = 'beverages',
  NATURE = 'nature',
  BODY_PARTS = 'bodyParts',
  BOOLEAN = 'boolean',
  BUILDINGS = 'buildings',
  CITIES = 'cities',
  CLOTHES = 'clothes',
  COLORS = 'colors',
  COMMUNICATION = 'communication',
  COMPARATIVE_WORD = 'comparativeWord',
  CONCEPTS = 'concepts',
  CONJUNCTIVE_WORD = 'conjunctiveWord',
  CUTLERY = 'cutlery',
  DAILY_ROUTINE = 'dailyRoutine',
  DAIRY_PRODUCTS = 'dairyProducts',
  DECLARATIVE_WORD = 'declarativeWord',
  EDUCATION = 'education',
  ENTERTAINMENT = 'entertainment',
  EXISTENCE = 'existence',
  FAMILY = 'family',
  FINANCIAL = 'financial',
  FOOD = 'food',
  FORMAL = 'formal',
  FRUITS = 'fruits',
  FURNITURE = 'furniture',
  GEOGRAPHICAL_NAMES = 'geographicalNames',
  GOVERNMENT_OFFICIAL = 'governmentOfficial',
  GREETINGS = 'greetings',
  HEALTH = 'health',
  INFORMAL = 'informal',
  JUSTICE = 'justice',
  LOCATIONS = 'locations',
  LUXURY = 'luxury',
  MATERIALS = 'materials',
  NEGATIVE_WORD = 'negativeWord',
  METALS = 'metals',
  MUSIC = 'music',
  NATIONALITY = 'nationality',
  NUMBERS = 'numbers',
  OBJECTS = 'objects',
  ONOMATOPOEIC_WORD = 'onomatopoeicWord',
  OUTER_SPACE = 'outerSpace',
  PERSONAL_NAMES = 'personalNames',
  PERSONAL_PRONOUN = 'personalPronoun',
  PLANTS = 'plants',
  PLURAL = 'plural',
  POLITICS = 'politics',
  POSITIVE_WORD = 'positiveWord',
  PROFESSION = 'profession',
  RELATIONS = 'relations',
  RIVERS = 'rivers',
  QUESTION_WORD = 'questionWord',
  SEASONS = 'seasons',
  SPICES = 'spices',
  SPORTS = 'sports',
  STUDYING = 'studying',
  SWEETS = 'sweets',
  TIME = 'time',
  TOOLS = 'tools',
  TRANSPORTATION = 'transportation',
  UNITS_OF_MEASUREMENT = 'unitsOfMeasurement',
  WEATHER = 'weather',
  WRITING = 'writing',
  SCIENCES = 'sciences',
}